Add Info storing

+++ client/src/commonMain/kotlin/io/bitpogo/pixalb/client/PixabayClient.kt
@@ -46,7 +46,7 @@ internal class PixabayClient internal constructor(
 
     private suspend fun fetchImagesFromApi(
         query: String,
-        page: UInt
+        page: UShort
     ): PixabayResponse {
         return requestBuilder
             .create()
@@ -64,7 +64,7 @@ internal class PixabayClient internal constructor(
 
     override suspend fun fetchImages(
         query: String,
-        page: UInt
+        page: UShort
     ): ResultContract<PixabayResponse, PixabayClientError> = guardTransaction {
         try {
             Success(fetchImagesFromApi(query, page))
